Replaced np.float

+++ id3/id3.py
@@ -107,7 +107,7 @@ class Id3Estimator(BaseEstimator):
             min_samples_split = 1
 
         if isinstance(self.min_entropy_decrease,
-                      (np.float, np.integer)):
+                      (float, np.integer)):
             min_entropy_decrease = (0 if self.min_entropy_decrease < 0
                                     else self.min_entropy_decrease)
         else:
